auth0によるユーザー認可処理追加

+++ src/pages/api/users/[id].ts
@@ -1,38 +1,51 @@
-import { NextRequest, NextResponse } from "next/server";
-import { PrismaClient } from "@prisma/client";
 import prisma from "../prisma";
 import { NextApiRequest, NextApiResponse } from "next";
+import { getSession, withApiAuthRequired } from "@auth0/nextjs-auth0";
 
-export default async function handle(
-  req: NextApiRequest,
-  res: NextApiResponse
-) {
-  const { id, accountId } = req.query;
-  console.log(req.query);
+const handle = async (req: NextApiRequest, res: NextApiResponse) => {
+  const { id: accountId } = req.query;
+  const { user: atuth0User } = await getSession(req, res);
+
   if (req.method === "GET") {
     let user;
-    if (id) {
-      user = await prisma.user.findUnique({
-        where: { id: String(id) },
-      });
-    } else if (accountId) {
+    if (accountId) {
       user = await prisma.user.findUnique({
         where: { accountId: String(accountId) },
       });
     }
     res.json(user);
   } else if (req.method === "PUT") {
-    const user = await prisma.user.update({
-      where: { id: String(id) },
-      data: req.body,
-    });
+    if (!atuth0User.sub || atuth0User.sub !== accountId) {
+      return res
+        .status(403)
+        .json({ error: "You can only access your own data." });
+    }
+
+    let user;
+    if (accountId) {
+      user = await prisma.user.update({
+        where: { accountId: String(accountId) },
+        data: req.body,
+      });
+    }
     res.json(user);
   } else if (req.method === "DELETE") {
-    const user = await prisma.user.delete({
-      where: { id: String(id) },
-    });
+    if (!atuth0User.sub || atuth0User.sub !== accountId) {
+      return res
+        .status(403)
+        .json({ error: "You can only access your own data." });
+    }
+
+    let user;
+    if (accountId) {
+      user = await prisma.user.delete({
+        where: { accountId: String(accountId) },
+      });
+    }
     res.json(user);
   } else {
     res.status(405).end();
   }
-}
+};
+
+export default withApiAuthRequired(handle);
